chore(project): refactor how properties are written to business objects
Since this commit it is not necessary to have a valid state of
properties in a business object. This allows us to merely change and
persist parts of complex properties.

Furthermore it is not possible to set invalid IDs to business objects,
but the user still gets feedback that his/her input is invalid.

Closes #133