feat(disable): provide a hook to disable a DOM element
The attribute [data-disable] can be used to disable a DOM element.

For example:

<input data-disable="callback">

Whenever an update of the DOM happens, the attribute [data-disable]
is evaluated. Therefore the given function is called, in that case
it is the function "callback()" of an entry. If the returned result
is true, then the attribute [disabled] is set to the DOM element.

As a result, if "callback" return true, the resulting DOM element
looks as follows:

<input data-disable="callback" disabled >