feat(callActivity/inOut): add camunda:In and camunda:Out parameters
Create a new tab 'Variables' to configure camunda:In and camunda:Out
parameters for a CallActivity. One of parameter source or sourceExpression
must have a value and then the target attribute is required.
To add/remove all variables for camunda:In and/or camunda:Out
enable/disable the checkboxes.

related to #137